+++ src/frontend/apps/standalone_site/src/features/Language/store/languageStore.spec.ts
@@ -1,17 +1,12 @@
 import { useLanguageStore } from './languageStore';
 
-jest.mock('../utils', () => {
-  return {
-    getLanguage: () => 'en_Test',
-  };
-});
-
 describe('stores/useLanguageStore', () => {
-  afterEach(() => {
-    jest.resetAllMocks();
+  it('checks the default language', () => {
+    expect(useLanguageStore.getState().language).toEqual('en_US');
   });
 
-  it('checks the default language', () => {
-    expect(useLanguageStore.getState().language).toEqual('en_Test');
+  it('sets another language', () => {
+    useLanguageStore.getState().setLanguage('fr');
+    expect(useLanguageStore.getState().language).toEqual('fr_FR');
   });
 });