+++ src/frontend/packages/lib_common/src/theme.ts
@@ -403,9 +403,6 @@ export const theme: ThemeType = {
           border-radius: 4px;
           margin: 0.125rem 0 0.125rem 1%;
         }
-        & button > div {
-          padding: 1rem;
-        }
         & button:hover {
           background-color: ${colorsGeneric['blue-hover-light']};
           color: inherit;
@@ -423,7 +420,7 @@ export const theme: ThemeType = {
     },
     options: {
       container: {
-        pad: 'small',
+        pad: '1rem',
       },
       text: {
         size: '0.875rem',